chore: Reduce snackbar duration
Changes the `SnackbarDuration` from `Long` to `Short` in `MainApp` to provide a quicker user feedback experience for showing messages.

+++ app/src/main/java/dev/alimansour/mytasks/MainApp.kt
@@ -51,7 +51,7 @@ fun MainApp() {
                 coroutineScope.launch {
                     snackbarHostState.showSnackbar(
                         message = uiText.asString(context),
-                        duration = SnackbarDuration.Long,
+                        duration = SnackbarDuration.Short,
                     )
                 }
             },
@@ -59,7 +59,7 @@ fun MainApp() {
                 coroutineScope.launch {
                     snackbarHostState.showSnackbar(
                         message = uiText.asString(context),
-                        duration = SnackbarDuration.Long,
+                        duration = SnackbarDuration.Short,
                     )
                 }
             },
